# docs/swagger.yaml
basePath: /api
definitions:
...
  request.CreatePartyRequest:
    properties:
      host_id:
        description: |-
          ID del usuario que será el anfitrión de la party
          example: 1
          required: true
        type: integer
      movie_id:
        description: |-
          ID de la película que se verá en la party
          example: 42
          required: true
        type: integer
    required:
    - host_id
    - movie_id
    type: object
...
  response.ErrorResponse:
    properties:
      error:
        description: |-
          Detalle técnico del error (opcional)
          example: "Invalid email format"
        type: string
      message:
        description: |-
          Mensaje descriptivo del error
          example: "Solicitud inválida"
        type: string
      statusCode:
        description: |-
          Código de estado HTTP
          example: 400
        type: integer
    type: object
...
  response.SuccessResponse:
    properties:
      data:
        description: Datos adicionales de la operación (opcional)
      message:
        description: |-
          Mensaje descriptivo del resultado
          example: Operación realizada con éxito
        type: string
      success:
        description: |-
          Indica si la operación fue exitosa
          example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: API para la plataforma de streaming de películas clásicas.
  title: Vintage Vision API
  version: "1.0"
paths:
...
  /api/party:
    post:
      consumes:
      - application/json
      description: Crea una nueva party con el host y película especificados
      parameters:
      - description: Datos para crear la party
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreatePartyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Party creada exitosamente
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Entrada inválida
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Error al crear la party
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Crear una nueva party
      tags:
      - Parties
  /api/party/join/{code}:
    post:
      consumes:
      - application/json
      description: Permite a un perfil unirse a una party usando su código
      parameters:
      - description: Código de la party
        in: path
        name: code
        required: true
        type: string
      - description: ID del perfil
        in: header
        name: profile_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Unión exitosa a la party
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Código de party o ID de perfil faltante
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Error al unirse a la party
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Unirse a una party existente
      tags:
      - Parties
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"